feat(projects): doing sidebar runs
