--- frontend/src/atoms/dialog.atoms.ts
@@ -1,4 +1,0 @@
-import { atom } from "jotai";
-import { atomFamily } from "jotai/utils";
-
-export const dialogOpenedAtomFamily = atomFamily(() => atom(false));
