--- frontend/src/atoms/navigation.atoms.ts
@@ -1,3 +1,0 @@
-import { atom } from "jotai";
-
-export const pageTitleAtom = atom("");
