+++ frontend/src/components/EmptyState.tsx
