+++ frontend/src/components/Header.tsx
@@ -1,4 +1,3 @@
-import { userAtom } from "@/atoms/user.atoms";
 import { Disclosure, Menu, Transition } from "@headlessui/react";
 import { Bars3Icon, XMarkIcon } from "@heroicons/react/24/outline";
 import clsx from "clsx";
@@ -7,6 +6,7 @@ import { BellIcon } from "lucide-react";
 import { Fragment } from "react";
 import { MobileNav, Nav } from "./Nav";
 import { RouteItem } from "./NavigationItem";
+import { userAtom } from "@/pages/atoms";
 
 export interface HeaderProps {
   routes: RouteItem[];
@@ -14,6 +14,7 @@ export interface HeaderProps {
 }
 
 export function Header({ routes, userMenu }: HeaderProps) {
+  /** @TODO This shoud be deleted and place at top level when the back end supports users */
   const user = useAtomValue(userAtom);
 
   return (
