+++ frontend/src/components/NavigationItem.tsx
