+++ frontend/src/components/ServerError.tsx
