+++ frontend/src/enums/dialog-ids.enum.ts
@@ -1,4 +1,3 @@
 export enum DialogIds {
   CreateService = "CreateService",
-  EditService = "EditService",
 }
