+++ frontend/src/pages/self-service/atoms.tsx
@@ -1,0 +1,1 @@
+export const SELF_SERVICE = "selfServiceSections";
