+++ frontend/src/pages/self-service/catalog-list/atoms.tsx
@@ -3,6 +3,7 @@ import { atomWithSwr } from "@/lib/atom-with-swr";
 import { makeQueryAtoms } from "@/lib/make-query-atoms";
 import { Catalog } from "@/types/catalog.type";
 import { atom } from "jotai";
+import { SELF_SERVICE } from "../atoms";
 
 export const selectedServiceSlugAtom = atom<string | null>(null);
 
@@ -13,12 +14,12 @@ export const [catalogsAtom, catalogsStatusAtom] = makeQueryAtoms<
   Catalog[]
 >(
   "catalogs",
-  () => `${API_URL}/catalogs`,
+  () => `${API_URL}/${SELF_SERVICE}`,
   (params) => () => {
     return fetch(params)
       .then((res) => res.json())
       .then((data) => data.results);
-  }
+  },
 );
 
 export const catalogsSwrAtom = atomWithSwr(catalogsAtom);
