+++ frontend/src/types/user.type.ts
@@ -1,0 +1,6 @@
+export type User = {
+  id: string;
+  name: string;
+  email: string;
+  imageUrl: string;
+};
